LayoutTests/imported/w3c/web-platform-tests/css/css-variables/revert-rule-in-fallback-expected.txt


FAIL var(--unknown, revert-rule) in custom property assert_equals: expected "PASS" but got "revert-rule"
FAIL var(--unknown, revert-rule) in shorthand assert_true: expected true got false
FAIL var(--unknown, revert-rule) in shorthand observed via longhand assert_true: expected true got false
FAIL var(--unknown, revert-rule) in longhand assert_true: expected true got false
